backend/requirements.txt
Flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
Flask-Cors==4.0.0
google-generativeai==0.3.2